feat: 짝꿍 만들기 1단계에서 2단계 넘어가는 기능 구현
